pyproject.toml
[project]
name = "garmin-mcp"
version = "0.1.0"
description = "MCP server giving an LLM direct access to a single user's Garmin Connect data."
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.0",
    "starlette>=0.37",
    "uvicorn[standard]>=0.30",
    "garminconnect>=0.2.20",
    "fitparse>=1.2.0",
    "google-cloud-storage>=2.16",
    "google-cloud-secret-manager>=2.20",
]

[project.scripts]
garmin-mcp = "garmin_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/garmin_mcp"]

[dependency-groups]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]
